Time class

+++ src/by/svetilnik/epam/d_classes/simpleClasses/task7/Top.java
@@ -1,27 +1,27 @@
 package by.svetilnik.epam.d_classes.simpleClasses.task7;
 
 public class Top {
-    private int oneParam;
-    private int twoParam;
+    private double oneParam;
+    private double twoParam;
 
-    public Top(int oneParam, int twoParam) {
+    public Top(double oneParam, double twoParam) {
         this.oneParam = oneParam;
         this.twoParam = twoParam;
     }
 
-    public int getOneParam() {
+    public double getOneParam() {
         return oneParam;
     }
 
-    public void setOneParam(int oneParam) {
+    public void setOneParam(double oneParam) {
         this.oneParam = oneParam;
     }
 
-    public int getTwoParam() {
+    public double getTwoParam() {
         return twoParam;
     }
 
-    public void setTwoParam(int twoParam) {
+    public void setTwoParam(double twoParam) {
         this.twoParam = twoParam;
     }
 }
